refactor(ui): extract persistence stores + autosave from main.js (S5 Stage 2)
Pure move, no behavior change:

- ui/electron-main/settings.js — user settings store (xleth-settings.json:
  loadSettings/saveSettings, sanitizeGlobalStretchMethod,
  getNewProjectGlobalStretchMethodDefault) plus the layout.json
  xleth:layout:read/write handlers that lived under the Themes header.
  xleth:settings:get/set stay in main.js — they are entangled with the
  workspace-backdrop logic, not with the store.
- ui/electron-main/autosave.js — autosave timer + xleth:autosave:restart.
  Worker access via ./worker, settings via ./settings; the export-progress
  gate stays in main.js and is probed via injected isExportBusy.
- ui/electron-main/themes.js — userData/themes/<slug>.json store.
- ui/electron-main/plugin-ui-layouts.js — userData/plugin-ui/<pluginId>.json
  store incl. xleth:dialog:importPluginUi/exportPluginUi; dialogs parent via
  injected getWin. Exports pluginIdSafe/validateLayoutStructure/
  parseImportedPluginUiLayout for other main-process callers.
- ui/electron-main/knob-presets.js — userData/plugin-ui-presets/knob.json.
- ui/electron-main/decals.js — userData/plugin-ui-assets/ store incl. magic
  checks and index self-healing; store primitives exported.
- ui/electron-main/grid-layout.js — engine pass-through grid handlers,
  registered inside init({ safeHandler }) since they wrap main.js's
  safeHandler at registration time.

Moved code is verbatim; every IPC channel name, userData file path, default
value and validation rule is unchanged. Handlers register at require time,
matching the module-scope timing they had in main.js. electron-builder.json
already whitelists electron-main/**/* (Stage 1), so no packaging change.

Global clip-processing defaults, Quick Launchers, and everything flagged
perf-sensitive in AUDIT.md §4.3 (WORLD poll, shm frame output) are untouched;
handler tables by domain remain for the next stage.

Verified: npm run build && npm run test:smoke -> 3/3 against both the working
tree and this exact committed main.js; vitest: 2478 passed, 2 failed = the
pre-existing local-only mixer failures (uncommitted FxBadge WIP), identical
to the Stage 1 baseline.
